diffusion/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.6.0=pypi_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.11.17=py38h06a4308_0
charset-normalizer=3.3.2=pypi_0
idna=3.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
ncurses=6.4=h6a678d5_0
numpy=1.24.4=pypi_0
openssl=3.0.12=h7f8727e_0
pandas=2.0.3=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py38h06a4308_0
pydantic=2.5.2=pypi_0
pydantic-core=2.14.5=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tornado=6.3.3=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.1.0=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
